--- app/assets/config/rails_ruby_lsp_manifest.js
@@ -1,1 +1,0 @@
-//= link_directory ../stylesheets/rails_ruby_lsp .css
