+++ app/assets/config/ruby_lsp_rails_manifest.js
@@ -1,0 +1,1 @@
+//= link_directory ../stylesheets/ruby_lsp_rails .css
