--- app/controllers/rails_ruby_lsp/application_controller.rb
@@ -1,7 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module RailsRubyLsp
-  class ApplicationController < ActionController::Base
-  end
-end
