+++ app/controllers/ruby_lsp/rails/application_controller.rb
@@ -1,0 +1,9 @@
+# typed: strict
+# frozen_string_literal: true
+
+module RubyLsp
+  module Rails
+    class ApplicationController < ActionController::Base
+    end
+  end
+end
