+++ app/jobs/ruby_lsp_rails/application_job.rb
@@ -1,7 +1,7 @@
 # typed: strict
 # frozen_string_literal: true
 
-module RailsRubyLsp
+module RubyLspRails
   class ApplicationJob < ActiveJob::Base
   end
 end
