+++ app/mailers/ruby_lsp_rails/application_mailer.rb
@@ -1,7 +1,7 @@
 # typed: strict
 # frozen_string_literal: true
 
-module RailsRubyLsp
+module RubyLspRails
   class ApplicationMailer < ActionMailer::Base
     default from: "from@example.com"
     layout "mailer"
