--- app/models/rails_ruby_lsp/application_record.rb
@@ -1,8 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module RailsRubyLsp
-  class ApplicationRecord < ActiveRecord::Base
-    self.abstract_class = true
-  end
-end
