+++ app/views/layouts/ruby_lsp_rails/application.html.erb
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <title>Rails ruby lsp</title>
+  <title>Ruby LSP Rails</title>
   <%= csrf_meta_tags %>
   <%= csp_meta_tag %>
 
-  <%= stylesheet_link_tag    "rails_ruby_lsp/application", media: "all" %>
+  <%= stylesheet_link_tag    "ruby_lsp_rails/application", media: "all" %>
 </head>
 <body>
