+++ bin/rails
@@ -5,7 +5,7 @@
 # installed from the root of your application.
 
 ENGINE_ROOT = File.expand_path("..", __dir__)
-ENGINE_PATH = File.expand_path("../lib/rails_ruby_lsp/engine", __dir__)
+ENGINE_PATH = File.expand_path("../lib/ruby_lsp_rails/engine", __dir__)
 APP_PATH = File.expand_path("../test/dummy/config/application", __dir__)
 
 # Set up gems listed in the Gemfile.
