+++ config/routes.rb
@@ -1,6 +1,6 @@
 # typed: strict
 # frozen_string_literal: true
 
-RailsRubyLsp::Engine.routes.draw do
+RubyLsp::Rails::Engine.routes.draw do
   resources :models, only: [:show]
 end
