+++ dev.yml
@@ -1,4 +1,4 @@
-name: rails-ruby-lsp
+name: ruby-lsp-rails
 
 type: ruby
