--- lib/rails_ruby_lsp.rb
@@ -1,9 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "sorbet-runtime"
-require "rails_ruby_lsp/version"
-require "rails_ruby_lsp/engine"
-
-module RailsRubyLsp
-end
