--- lib/rails_ruby_lsp/version.rb
@@ -1,6 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module RailsRubyLsp
-  VERSION = "0.1.0"
-end
