+++ lib/ruby-lsp-rails.rb
@@ -1,0 +1,11 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "sorbet-runtime"
+require "ruby_lsp_rails/version"
+require "ruby_lsp_rails/engine"
+
+module RubyLSP
+  module Rails
+  end
+end
