+++ lib/ruby_lsp_rails/version.rb
@@ -1,7 +1,8 @@
 # typed: strict
 # frozen_string_literal: true
 
-module RailsRubyLsp
-  module ApplicationHelper
+module RubyLsp
+  module Rails
+    VERSION = "0.1.0"
   end
 end
