+++ lib/tasks/ruby_lsp_rails_tasks.rake
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 # desc "Explaining what the task does"
-# task :rails_ruby_lsp do
+# task :ruby_lsp_rails do
 #   # Task goes here
 # end
