+++ sorbet/rbi/dsl/generated_path_helpers_module.rbi
@@ -78,9 +78,6 @@ module GeneratedPathHelpersModule
   def rails_representation_path(*args); end
 
   sig { params(args: T.untyped).returns(String) }
-  def rails_ruby_lsp_path(*args); end
-
-  sig { params(args: T.untyped).returns(String) }
   def rails_sendgrid_inbound_emails_path(*args); end
 
   sig { params(args: T.untyped).returns(String) }
@@ -94,6 +91,9 @@ module GeneratedPathHelpersModule
 
   sig { params(args: T.untyped).returns(String) }
   def rails_storage_redirect_path(*args); end
+
+  sig { params(args: T.untyped).returns(String) }
+  def ruby_lsp_rails_path(*args); end
 
   sig { params(args: T.untyped).returns(String) }
   def update_rails_disk_service_path(*args); end
