--- sorbet/rbi/dsl/rails_ruby_lsp/application_mailer.rbi
@@ -1,7 +1,0 @@
-# typed: true
-
-# DO NOT EDIT MANUALLY
-# This is an autogenerated file for dynamic methods in `RailsRubyLsp::ApplicationMailer`.
-# Please instead update this file by running `bin/tapioca dsl RailsRubyLsp::ApplicationMailer`.
-
-class RailsRubyLsp::ApplicationMailer; end
