+++ test/dummy/config/application.rb
@@ -8,7 +8,7 @@ require "rails/all"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-require "rails_ruby_lsp"
+require "ruby-lsp-rails"
 
 module Dummy
   class Application < Rails::Application
