--- test/rails_ruby_lsp_test.rb
@@ -1,10 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-require "test_helper"
-
-class RailsRubyLspTest < ActiveSupport::TestCase
-  test "it has a version number" do
-    assert RailsRubyLsp::VERSION
-  end
-end
