+++ test/ruby_lsp_rails_test.rb
@@ -1,0 +1,12 @@
+# typed: true
+# frozen_string_literal: true
+
+require "test_helper"
+
+module RubyLsp
+  class RailsTest < ActiveSupport::TestCase
+    test "it has a version number" do
+      assert RubyLsp::Rails::VERSION
+    end
+  end
+end
